ZF 1.10 to Zend 1.20(ZF1 future)

+++ library/Zend/Captcha/Image.php
@@ -15,9 +15,9 @@
  * @category   Zend
  * @package    Zend_Captcha
  * @subpackage Adapter
- * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright  Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
- * @version    $Id: Image.php 20096 2010-01-06 02:05:09Z bkarwin $
+ * @version    $Id$
  */
 
 /** @see Zend_Captcha_Word */
@@ -31,7 +31,7 @@ require_once 'Zend/Captcha/Word.php';
  * @category   Zend
  * @package    Zend_Captcha
  * @subpackage Adapter
- * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright  Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
 class Zend_Captcha_Image extends Zend_Captcha_Word
@@ -240,24 +240,36 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
     {
         return $this->_width;
     }
+
     /**
+     * Set start image
+     *
      * @param string $startImage
+     * @return Zend_Captcha_Image
      */
     public function setStartImage ($startImage)
     {
         $this->_startImage = $startImage;
         return $this;
     }
+
     /**
+     * Set dot noise level
+     *
      * @param int $dotNoiseLevel
+     * @return Zend_Captcha_Image
      */
     public function setDotNoiseLevel ($dotNoiseLevel)
     {
         $this->_dotNoiseLevel = $dotNoiseLevel;
         return $this;
     }
-   /**
+
+    /**
+     * Set line noise level
+     *
      * @param int $lineNoiseLevel
+     * @return Zend_Captcha_Image
      */
     public function setLineNoiseLevel ($lineNoiseLevel)
     {
@@ -348,8 +360,12 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
         $this->_imgUrl = rtrim($imgUrl, "/\\") . '/';
         return $this;
     }
+
     /**
+     * Set image alternative text
+     *
      * @param string $imgAlt
+     * @return Zend_Captcha_Image
      */
     public function setImgAlt ($imgAlt)
     {
@@ -426,11 +442,13 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
             $id = $this->_generateRandomId();
             $this->_setId($id);
         }
+
         $this->_generateImage($id, $this->getWord());
 
-        if (mt_rand(1, $this->getGcFreq()) == 1) {
+        if (mt_rand(1, $this->getGcFreq()) === 1) {
             $this->_gc();
         }
+
         return $id;
     }
 
@@ -440,8 +458,9 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
      * Override this function if you want different image generator
      * Wave transform from http://www.captcha.ru/captchas/multiwave/
      *
-     * @param string $id Captcha ID
+     * @param string $id   Captcha ID
      * @param string $word Captcha word
+     * @throws Zend_Captcha_Exception
      */
     protected function _generateImage($id, $word)
     {
@@ -524,16 +543,19 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
 
                 if ($sx < 0 || $sy < 0 || $sx >= $w - 1 || $sy >= $h - 1) {
                     continue;
-                } else {
-                    $color    = (imagecolorat($img, $sx, $sy) >> 16)         & 0xFF;
-                    $color_x  = (imagecolorat($img, $sx + 1, $sy) >> 16)     & 0xFF;
-                    $color_y  = (imagecolorat($img, $sx, $sy + 1) >> 16)     & 0xFF;
-                    $color_xy = (imagecolorat($img, $sx + 1, $sy + 1) >> 16) & 0xFF;
                 }
+
+                $color    = (imagecolorat($img, $sx, $sy) >> 16)         & 0xFF;
+                $color_x  = (imagecolorat($img, $sx + 1, $sy) >> 16)     & 0xFF;
+                $color_y  = (imagecolorat($img, $sx, $sy + 1) >> 16)     & 0xFF;
+                $color_xy = (imagecolorat($img, $sx + 1, $sy + 1) >> 16) & 0xFF;
+
                 if ($color == 255 && $color_x == 255 && $color_y == 255 && $color_xy == 255) {
                     // ignore background
                     continue;
-                } elseif ($color == 0 && $color_x == 0 && $color_y == 0 && $color_xy == 0) {
+                }
+
+                if ($color == 0 && $color_x == 0 && $color_y == 0 && $color_xy == 0) {
                     // transfer inside of the image as-is
                     $newcolor = 0;
                 } else {
@@ -567,7 +589,6 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
 
     /**
      * Remove old files from image directory
-     *
      */
     protected function _gc()
     {
@@ -577,10 +598,14 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
             // safety guard
             return;
         }
+        $suffixLength = strlen($this->_suffix);
         foreach (new DirectoryIterator($imgdir) as $file) {
             if (!$file->isDot() && !$file->isDir()) {
-                if ($file->getMTime() < $expire) {
-                    unlink($file->getPathname());
+                if (file_exists($file->getPathname()) && $file->getMTime() < $expire) {
+                    // only deletes files ending with $this->_suffix
+                    if (substr($file->getFilename(), -($suffixLength)) == $this->_suffix) {
+                        unlink($file->getPathname());
+                    }
                 }
             }
         }
@@ -595,6 +620,11 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
      */
     public function render(Zend_View_Interface $view = null, $element = null)
     {
-        return '<img width="'.$this->getWidth().'" height="'.$this->getHeight().'" alt="'.$this->getImgAlt().'" src="' . $this->getImgUrl() . $this->getId() . $this->getSuffix() . '"/><br/>';
+        $endTag = ' />';
+        if (($view instanceof Zend_View_Abstract) && !$view->doctype()->isXhtml()) {
+            $endTag = '>';
+        }
+        return '<img width="' . $this->getWidth() . '" height="' . $this->getHeight() . '" alt="' . $this->getImgAlt()
+             . '" src="' . $this->getImgUrl() . $this->getId() . $this->getSuffix() . '"' . $endTag;
     }
 }
